removing the gradient adding the filter button
